Update.
	* include/bits/xopen_lim.h (LONG_BIT): Use LONG_MAX, not INT_MAX.
	Patch by Thorsten Kukuk <kukuk@suse.de>.

+++ include/bits/xopen_lim.h
@@ -104,7 +104,7 @@
 #endif
 
 /* Number of bits in a word of type `long int'.  */
-#if INT_MAX == 2147483647
+#if LONG_MAX == 2147483647
 # define LONG_BIT	32
 #else
 /* Safe assumption.  */
